+++ parametricSN/training/cross_entropy_training.py
@@ -12,7 +12,6 @@ from sklearn.metrics import classification_report
 
 def test(model, device, test_loader):
     """test method"""
-
     model.eval()
     test_loss = 0
     correct = 0
@@ -20,7 +19,9 @@ def test(model, device, test_loader):
     with torch.no_grad():
         for data, target in test_loader:
             data, target = data.to(device), target.to(device, dtype=torch.long)  
+            print("fsdhwakhkDS")
             output = model(data)
+            print("FSFQWQ")
             test_loss += F.cross_entropy(output, target, reduction='sum').item() # sum up batch loss
             pred = output.max(1, keepdim=True)[1] # get the index of the max log-probability
             correct += pred.eq(target.view_as(pred)).sum().item()
@@ -33,7 +34,6 @@ def test(model, device, test_loader):
 
 def train(model, device, train_loader, scheduler, optimizer, epoch, accum_step_multiple=None):
     """training method"""
-
     model.train()
     correct = 0
     train_loss = 0
@@ -43,16 +43,13 @@ def train(model, device, train_loader, scheduler, optimizer, epoch, accum_step_m
 
         optimizer.zero_grad()
         output = model(data)
+        print("DS", batch_idx)
         loss = F.cross_entropy(output, target)
         loss.backward()
-
-        model.scatteringBase.saveFilterGrads(scatteringActive=True) 
         optimizer.step()
-        model.scatteringBase.saveFilterValues(scatteringActive=True) 
 
         if scheduler != None:
             scheduler.step()
-
 
         pred = output.max(1, keepdim=True)[1] # get the index of the max log-probabilityd
         correct += pred.eq(target.view_as(pred)).sum().item()
@@ -60,7 +57,6 @@ def train(model, device, train_loader, scheduler, optimizer, epoch, accum_step_m
     
     train_loss /= len(train_loader.dataset)
     train_accuracy = 100. * correct / len(train_loader.dataset)
-
     
     print('[Model -- {}] Train Epoch: {:>6} Average Loss: {:.6f}, Accuracy: {}/{} ({:.2f}%)'.format(
             model, epoch, train_loss, correct,
